+++ src/neuroconv/datainterfaces/__init__.py
@@ -56,6 +56,7 @@ from .ecephys.phy.phydatainterface import PhySortingInterface
 from .ecephys.plexon.plexondatainterface import (
     Plexon2RecordingInterface,
     PlexonRecordingInterface,
+    PlexonLFPInterface,
     PlexonSortingInterface,
 )
 from .ecephys.spike2.spike2datainterface import Spike2RecordingInterface
@@ -129,6 +130,7 @@ interface_list = [
     EDFRecordingInterface,
     TdtRecordingInterface,
     PlexonRecordingInterface,
+    PlexonLFPInterface,
     Plexon2RecordingInterface,
     PlexonSortingInterface,
     BiocamRecordingInterface,